+++ src/components/three/utils/borderVerticesUtils.js
@@ -1,0 +1,28 @@
+import * as THREE from 'three';
+
+const interpolatePoints = (start, end, numPoints) => {
+  const points = [];
+
+  for (let i = 0; i <= numPoints; i++) {
+    const t = i / numPoints;
+    const x = start.x + t * (end.x - start.x);
+    const y = start.y + t * (end.y - start.y);
+    const z = start.z + t * (end.z - start.z);
+    points.push(new THREE.Vector3(x, y, z));
+  }
+
+  return points;
+};
+
+export const generateBorderPoints = (corners, pointsPerEdge = 9) => {
+  const borderPoints = [];
+
+  for (let i = 0; i < corners.length; i++) {
+    const start = corners[i];
+    const end = corners[(i + 1) % corners.length];
+    const interpolatedPoints = interpolatePoints(start, end, pointsPerEdge);
+    borderPoints.push(...interpolatedPoints);
+  }
+
+  return borderPoints;
+};
